- add support for diff abstract assembly
