+++ AbsInjectTypeDll/AbsInjectTypeDll.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ShareAttributes\ShareAttributes.csproj" />
+  </ItemGroup>
+
+</Project>
